requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
cryptography
jinja2
python-multipart
python-jose
passlib 
bcrypt==4.0.1
pydantic
itsdangerous
psycopg2-binary
python-dotenv
authlib
httpx
